[GEOT-7422] fork Indriya's `SimpleUnitFormat` as `BaseUnitFormatter` and make existing classes us it
The `BaseUnitFormatter` is immutable and does not operate on global state.
Instead, all relevant configuration happens during construction through
use-case specific subclasses.
