+++ modules/library/metadata/src/main/java/org/geotools/measure/UnitFormatter.java
@@ -18,20 +18,22 @@ package org.geotools.measure;
 
 import java.io.IOException;
 import java.text.ParsePosition;
+import javax.measure.MeasurementException;
 import javax.measure.Unit;
 import javax.measure.format.MeasurementParseException;
 
-/**
- * An interface that is similar to {@link javax.measure.format.UnitFormat} but elides mutating
- * methods.
- *
- * <p>It is used to protect global or shared UnitFormat instances from being changed inadvertently.
- */
+/** An interface similar to {@link javax.measure.format.UnitFormat} but without mutating methods. */
 public interface UnitFormatter {
 
     Appendable format(Unit<?> unit, Appendable appendable) throws IOException;
 
-    String format(Unit<?> unit);
+    default String format(Unit<?> unit) {
+        try {
+            return this.format(unit, new StringBuilder()).toString();
+        } catch (IOException ex) {
+            throw new MeasurementException(ex); // Should never happen.
+        }
+    }
 
     Unit<?> parse(CharSequence csq, ParsePosition pos)
             throws IllegalArgumentException, MeasurementParseException;
