+++ modules/library/referencing/src/main/java/org/geotools/referencing/wkt/EsriUnitFormat.java
@@ -23,16 +23,17 @@ import java.util.stream.Stream;
 import org.geotools.measure.BaseUnitFormatter;
 import org.geotools.measure.UnitDefinition;
 import org.geotools.measure.UnitDefinitions;
-import org.geotools.measure.UnitFormatter;
 
 /** A factory for unit formatters that support the EPSG dialect. */
-public final class EsriUnitFormat {
+public final class EsriUnitFormat extends BaseUnitFormatter {
 
-    public static UnitFormatter getInstance() {
+    public static EsriUnitFormat getInstance() {
         return INSTANCE;
     }
 
-    private EsriUnitFormat() {}
+    private EsriUnitFormat(List<UnitDefinition> unitDefinitions) {
+        super(unitDefinitions);
+    }
 
     private static final List<UnitDefinition> UNIT_DEFINITIONS =
             Stream.of(
@@ -44,7 +45,7 @@ public final class EsriUnitFormat {
                             UnitDefinitions.US_CUSTOMARY,
                             UnitDefinitions.ESRI)
                     .flatMap(Collection::stream)
-                    .collect(Collectors.toList());
+                    .collect(Collectors.toUnmodifiableList());
 
-    private static final BaseUnitFormatter INSTANCE = new BaseUnitFormatter(UNIT_DEFINITIONS);
+    private static final EsriUnitFormat INSTANCE = new EsriUnitFormat(UNIT_DEFINITIONS);
 }
